fix CI, take two
