+++ label_studio/tests/ml/test_predict.py
@@ -2,63 +2,7 @@ import json
 
 import pytest
 
-from label_studio.tests.utils import make_project, make_task, register_ml_backend_mock
-
-
-@pytest.fixture
-def ml_backend_for_test_predict(ml_backend):
-    # ML backend with single prediction per task
-    register_ml_backend_mock(
-        ml_backend,
-        url='http://localhost:9092',
-        predictions={
-            'results': [
-                {
-                    'model_version': 'ModelSingle',
-                    'score': 0.1,
-                    'result': [
-                        {'from_name': 'label', 'to_name': 'text', 'type': 'choices', 'value': {'choices': ['Single']}}
-                    ],
-                },
-            ]
-        },
-    )
-    # ML backend with multiple predictions per task
-    register_ml_backend_mock(
-        ml_backend,
-        url='http://localhost:9093',
-        predictions={
-            'results': [
-                [
-                    {
-                        'model_version': 'ModelA',
-                        'score': 0.2,
-                        'result': [
-                            {
-                                'from_name': 'label',
-                                'to_name': 'text',
-                                'type': 'choices',
-                                'value': {'choices': ['label_A']},
-                            }
-                        ],
-                    },
-                    {
-                        'model_version': 'ModelB',
-                        'score': 0.3,
-                        'result': [
-                            {
-                                'from_name': 'label',
-                                'to_name': 'text',
-                                'type': 'choices',
-                                'value': {'choices': ['label_B']},
-                            }
-                        ],
-                    },
-                ]
-            ]
-        },
-    )
-    yield ml_backend
+from label_studio.tests.utils import make_project, make_task
 
 
 @pytest.mark.django_db
